chore: update .gitignore files to include .env.example and add dotenv dependency
