+++ backend/.env.example
@@ -1,0 +1,16 @@
+# Copy to .env and fill in real values. Never commit .env.
+
+# --- Admin auth (POST /admin/auth/login) ---
+# Plaintext comparison is OK for a single shared password in dev;
+# for production prefer a strong secret + proper user store or hashed password.
+ADMIN_PASSWORD=change-me
+
+# If you issue JWTs after login:
+JWT_SECRET=change-me-to-long-random-string
+JWT_EXPIRE_MINUTES=60
+
+# --- Quiz / existing flags ---
+QUIZ_SHOW_ANSWER_REVIEW=true
+
+# --- CORS (comma-separated); omit to use dev defaults (localhost:3000 + 127.0.0.1:3000) ---
+# CORS_ORIGINS=http://localhost:3000,http://127.0.0.1:3000
