+++ backend/pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.14"
 dependencies = [
     "fastapi>=0.135.3",
     "pydantic>=2.12.5",
+    "python-dotenv>=1.0.0",
     "uvicorn[standard]>=0.38.0",
 ]
