+++ backend/routes/nick.py
@@ -1,0 +1,12 @@
+"""Nickname validation."""
+
+from fastapi import APIRouter
+
+from routes.stubs import unimplemented
+
+router = APIRouter(tags=["nick"])
+
+
+@router.post("/validate-nick")
+async def validate_nick() -> None:
+    unimplemented()
