+++ backend/schemas/admin_auth.py
@@ -1,0 +1,1 @@
+"""Admin auth schemas (login/logout, tokens) — add when implementing endpoints."""
