+++ backend/schemas/admin_quiz.py
@@ -1,0 +1,1 @@
+"""Admin quiz API schemas (create/update, list, detail, leaderboard) — add when implementing endpoints."""
