+++ backend/schemas/common.py
@@ -1,0 +1,1 @@
+"""Shared models (errors, pagination, etc.) — add when needed."""
