+++ backend/schemas/nick.py
@@ -1,0 +1,1 @@
+"""Nickname validation schemas — add when implementing POST /validate-nick."""
