+++ frontend/.gitignore
@@ -30,8 +30,10 @@ yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
 
-# env files (can opt-in for committing if needed)
-.env*
+# env files (keep .env.example committable)
+.env
+.env.*
+!.env.example
 
 # vercel
 .vercel
